options: Fix multiple -v by simplifying CONF_TYPE_FUNC
The CONF_TYPE_FUNC implementation counted how many times the option
had appeared and called the function that many times when it was set.
However each -v on the command line triggered an increase of the count
_and_ as many calls as the current count indicated. Thus the resulting
verbosity levels for 0 to 5 -v were really 0, 1, 3, 6, 10, 15 instead
of 0, 1, 2, 3, 4, 5.

Remove the counting functionality and just call the given function
once at each set operation. This fixes -v; other options are not
affected.