Fix crash during early exit
Starting MPlayer with options such as "-demuxer help" that exit before
initializing OSD struct would crash at exit because of a reference to
mpctx->osd->sub_font in deinitialization code. Fix by checking that
mpctx->osd is not NULL.

Checking that osd->sub_font is not NULL was unnecessary because
free_font_desc(NULL) is safe.