demux: Reset demux stream 'eof' flag after packet buffer overflow
The eof flag is set when another demuxer stream has hit the limit on
the size or count of buffered packets. This makes sense because in that
error situation the calling code can not rely on being able to demux
more packets from this stream, so it should rather exit or whatever
instead of getting stuck trying. However the situation can improve if
packets are demuxed from the other stream. In that case the eof flag
should be cleared. This commit adds code to clear the flag if
something is successfully read from the stream (so it's only cleared
if a caller tries to read packets despite the flag being set; that's
enough to fix audio sync issues after video packet buffer overflow).