demux: Increase max buffered packet bytes from 8 to 128 MiB
The hard limit on the amount of buffered bytes per demuxer stream had
not been increased since 2001. Since then there have been significant
increases in both video bitrates (so it's easier to hit the limit) and
typical memory available on a computer (so there's less reason to
limit the memory used for buffering). The 8 MiB limit was too easy to
hit in the case of high-bitrate video lagging behind audio, when the
file is demuxed ahead to get audio packets but video has to be
buffered until the decoder catches up to that point.

Increase the limit to 128 MiB and remove the #ifdef for
CONFIG_TV_BSDBT848 because the normal limit is now higher than the
increased value under the #ifdef.