libass: If a bad subtitle has multiple \pos use first
Having multiple \pos makes no sense, but such files exist. Apparently
VSFilter uses the values from the first \pos in this case whereas
libass used the last one. Change the behavior to make it compatible
with VSfilter since this has the highest chance of matching what the
creators of the file intended.