core: Handle missing timestamp errors a bit better
If there is no timestamp available for a video frame at all, use
timestamp from the previous frame (if any) instead of leaving it as
MP_NOPTS_VALUE. The main practical difference for playback is that
MP_NOPTS_VALUE caused timing to be reset, moving current video
position immediately to time of the next frame that had a timestamp
and thus likely advancing the video too fast. Now the next frame with
a timestamp will be shown at its proper position relative to the
previous timestamp.