Fix ordered chapter pruning of redundant timeline parts
Code that was supposed to merge two timeline parts if the second one
started exactly where the first one stopped didn't work because it
mixed timestamps in different units (I had changed the units of some
variables after originally writing that code but forgotten to update
it). As a result of the bug there were unnecessary part switches, but
it was unlikely to cause any serious problems.