Revert "fix missing event on move that breaks xmga window movement"
The reverted commit had a harmful effect on other VOs as it produced
bogus events that caused unnecessary load and flicker at least when
moving the window while paused. A proper fix would need a separate
event type. So revert this commit before merging the changes to the
master branch; it'll leave xmga window movement broken, but I consider
that less important than the other VOs.