demux_mkv: Round more seeks to "backwards" keyframes
Allow more fuzziness in treating timestamps as inaccurate attempts to
seek "exactly" to a keyframe. A seek forward can now actually go to a
keyframe 5 ms back from the timecode, and correspondingly for seeks
backwards. Before only 1 ms was allowed. This helps with some
files that have inaccurate chapter boundaries.