Add copies of needed internal FFmpeg files under ffmpeg_files/
MPlayer used to depend on having an FFmpeg tree available at build
time to provide this code, and FFmpeg changes occasionally broke
the MPlayer build as a result. Some of the relevant functionality also
depends on FFmpeg library symbols that are not part of the FFmpeg API,
meaning that library changes could break MPlayer after the build. Add
a copy of the relevant functionality to the MPlayer tree and use that
instead.

The added files are:
- The headers bswap.h, intreadwrite.h and x86_cpu.h from libavutil.
  Some MPlayer code uses the functionality defined as macros and
  static inline functions in these headers.
- Architecture-specific internal files used by the above headers.
  These are in the subdirectories: x86/, arm/, bfin/, sh4/.
- taglists.[ch], which provides the codec tables from
  libavformat/riff.c and two functions to access them from
  libavformat/utils.c.

Most of the headers are copied from FFmpeg revision 19452 with only
the include guard names changed. A "common.h" include in bswap.h was
changed to "libavutil/common.h" as it should use the installed header.
Taglists.c contains snippets from the relevant files with some
changes like renamed identifiers.