vf_uspp: Mark as depending on avcodec API only, not internals
There has been some confusion about whether this filter depends on
libavcodec internals or not. When originally added it included
explicit code to support building with a shared library avcodec, but
it also used the internal dsputil.h header. Later it was marked as
depending on libavcodec internals, then the dsputil.h include was
removed as unnecessary. It seems to build fine without any internal
headers now and has no obvious other usage of internal symbols either,
so I'm enabling the build whenever libavcodec is available.