command.c: Split property OSD display from property commands
The function set_property_command() handles commands that require only
changing a property and then possibly displaying the new value on the
OSD somehow. Split it into two parts: one that executes commands by
doing property manipulation according to a table, and one that
displays the value of a given property on the OSD in a manner listed
for that property in another table. The goal is to allow using the
OSD-display code and information in generic commands that take the
property name as an argument.

Splitting the table in two was not strictly necessary - I could have
kept it as a single table with possibly dummy field values for entries
that only make sense for one part of the split function or the other.
However I think splitting it is cleaner and more readable.