vo_gl: Remove incorrect glXSwapIntervalEXT use
The list of OpenGL extension function names to try for setting the
swapinterval value contained "glXSwapIntervalEXT". Such a function
exists in the extension GL_EXT_swap_control but takes different
arguments than vo_gl expects, so trying to use it from the current
code would lead to unpredictable behavior. Remove it from the list of
functions to use.