ao_alsa: Reinitialize parameters properly when reopening
ao_alsa used static variables to define the parameter values for
snd_pcm_hw_params_set_buffer_time_near() and
snd_pcm_hw_params_set_periods_near(). The variables were non-const and
the desired value was only set in the initializer. The ALSA functions
in question take a pointer argument to an in/out parameter which is
modified to reflect the value actually used. As a result, when playing
multiple files or otherwise reinitializing the AO the later instances
could use values that had been modified by earlier calls. Change the
code to always always use the same default values.