vo_vdpau: improve VOCTRL_GET_EOSD_RES code
Remove the special case for fullscreen; the same variables should have
the correct values in windowed and fullscreen modes. This fixes libass
subtitle positioning in the (non-typical) case where you have black
bands inside a window. Behavior in fullscreen and in the typical
windowed case without borders should not be affected.