subs: support loading external ASS subtitles via stream layer
Previously the argument of the "-sub" option was always interpreted as
a local filename when trying to read it as a libass file. Use the
stream layer to read it instead, so that protocols like 'http://' are
also supported like they are for the main video/audio file.

Based on patch by Yuriy Kaminskiy <yumkam@mail.ru>